tested

+++ daft/runners/flotilla.py
@@ -174,7 +174,7 @@ def start_ray_workers(existing_worker_ids: list[str]) -> list[RaySwordfishWorker
     return handles
 
 
-def try_autoscale(bundles: list[dict[str, float]]) -> None:
+def try_autoscale(bundles: list[dict[str, int]]) -> None:
     from ray.autoscaler.sdk import request_resources
 
     request_resources(
